requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
Django==3.2.5
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==3.0.1
django-jquery==3.1.0
psycopg2-binary==2.9.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
